Rename calico_docker to calico_containers
In anticipation of a probable renaming of the repo.
