+++ calico_containers/adapter/docker_restart.py
