+++ calico_containers/adapter/netns.py
