+++ calico_containers/adapter/powerstrip_calico.py
