+++ calico_containers/adapter/requirements.txt
