+++ calico_containers/tests/fv/no_powerstrip.sh
