+++ calico_containers/tests/scale/cloud-config-master.yaml
