+++ calico_containers/tests/scale/cloud-config.yaml
