+++ calico_containers/tests/scale/kill.sh
