+++ calico_containers/tests/unit/datastore_test.py
@@ -2,7 +2,7 @@ __author__ = 'spike@projectcalico.org'
 
 
 from mock import patch, Mock, call
-from calico_docker.adapter.datastore import (DatastoreClient,
+from calico_containers.adapter.datastore import (DatastoreClient,
                                              Rule,
                                              Profile,
                                              Rules,
@@ -212,8 +212,8 @@ class TestEndpoint(unittest.TestCase):
 
 class TestDatastoreClient(unittest.TestCase):
 
-    @patch("calico_docker.adapter.datastore.os.getenv", autospec=True)
-    @patch("calico_docker.adapter.datastore.etcd.Client", autospec=True)
+    @patch("calico_containers.adapter.datastore.os.getenv", autospec=True)
+    @patch("calico_containers.adapter.datastore.etcd.Client", autospec=True)
     def setUp(self, m_etcd_client, m_getenv):
         m_getenv.return_value = "127.0.0.2:4002"
         self.etcd_client = Mock(spec=EtcdClient)
