+++ node_fs/etc/service/powerstrip-calico/run
@@ -1,3 +1,3 @@
 #!/bin/sh
-cd /calico_docker/adapter/
+cd /calico_containers/adapter/
 exec python powerstrip_calico.py
