+++ nose.cfg
@@ -2,5 +2,5 @@
 verbosity=3
 with-coverage=1
 cover-erase=1
-cover-package=calico_docker
+cover-package=calico_containers
 cover-html=1
